requirements.txt
Django
asgiref
autopep8
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
Pillow
django-mptt
